+++ docs/REQUIREMENTS.md
@@ -39,7 +39,7 @@
   - [x] MSW를 활용하여 실제 서버와 연동될 수 있는 API Mocking을 구현
   - [x] 단순한 Endpoint 변경으로 실제 API 사용이 가능하도록 작업
 - 테스트
-  - [ ] 장바구니 페이지에서 다양한 사용자 인터렉션에 대한 테스트 케이스를 고민하고, 선택한 테스트 도구를 이용하여 검증
+  - [x] 장바구니 페이지에서 다양한 사용자 인터렉션에 대한 테스트 케이스를 고민하고, 선택한 테스트 도구를 이용하여 검증
 - 사용자 경험
   - [x] 새로고침 해도, 장바구니에 담은 상품 유지
   - [x] 장바구니 페이지에서 개수가 1일 때 `-` 막기
