+++ src/components/ProductCardList/ProductCardList.tsx
@@ -1,6 +1,6 @@
 import { useRecoilValue } from 'recoil';
 import { styled } from 'styled-components';
-import { fetchedProductListSelector } from '../../store/fetchSelectors';
+import { fetchedProductListSelector } from '../../store/asyncSelector';
 import { WIDTH } from '../../styles/mediaQuery';
 import ProductCard from '../ProductCard/ProductCard';
